исправлены замечания по поводу добавления геттеров и тестов
